+++ Runtime/Vector2IntExtensions.cs
@@ -8,6 +8,16 @@ namespace UnityExtensions
     public static class Vector2IntExtensions
     {
         /// <summary>
+        /// Sets each component to its absolute value.
+        /// </summary>
+        /// <param name="vector">Vector.</param>
+        /// <returns>Vector with absolute values.</returns>
+        public static Vector2Int Abs(this Vector2Int vector)
+        {
+            return new Vector2Int(Mathf.Abs(vector.x), Mathf.Abs(vector.y));
+        }
+
+        /// <summary>
         /// Sets the x component of the vector.
         /// </summary>
         /// <param name="vector">Vector.</param>