+++ Runtime/Vector3IntExtensions.cs
@@ -8,6 +8,16 @@ namespace UnityExtensions
     public static class Vector3IntExtensions
     {
         /// <summary>
+        /// Sets each component to its absolute value.
+        /// </summary>
+        /// <param name="vector">Vector.</param>
+        /// <returns>Vector with absolute values.</returns>
+        public static Vector3Int Abs(this Vector3Int vector)
+        {
+            return new Vector3Int(Mathf.Abs(vector.x), Mathf.Abs(vector.y), Mathf.Abs(vector.z));
+        }
+
+        /// <summary>
         /// Sets the x component of the vector.
         /// </summary>
         /// <param name="vector">Vector.</param>